Fix compile errors/warnings in beep.c and compare.c (Closes: #1754218).
git-svn-id: https://secure.a-eskwadraat.nl/svn/domjudge/trunk@1556 caf94f87-7bd6-0310-88dc-dee8202e2269

+++ judge/compare.c
@@ -221,7 +221,7 @@ void writediff()
 	/* Determine left/right printing length and construct format
 	   string for printf later */
 	for(i=0; i<2; i++) maxlinelen[0] = min(maxlinelen[0],maxprintlen);
-	sprintf(formatstr,"%%3d '%%-%ds %%c '%%s\n",maxlinelen[0]+1);
+	sprintf(formatstr,"%%3d '%%-%ds %%c '%%s\n",(int)maxlinelen[0]+1);
 	
 	/* Loop over all common lines to detect differences */
 	for(l=0; l<min(nlines[0],nlines[1]); l++) {
